#355: 도메인별 설계 통합 및 파일 정리
