+++ Todoary-iOS/Todoary/Network/Entity/Diary/DiaryModel.swift
@@ -1,5 +1,5 @@
 //
-//  DiaryStickerResultModel.swift
+//  DiaryModel.swift
 //  Todoary
 //
 //  Created by 송채영 on 2023/01/12.
@@ -7,6 +7,20 @@
 
 import Foundation
 
+//MARK: - Text
+struct DiaryRequestModel: Codable{
+    let title: String
+    let content: String
+}
+
+struct DiaryResultModel: Codable{
+    let diaryId: Int
+    var title: String
+    var content: String
+    let created_at: String
+}
+
+//MARK: - Sticker
 struct DiaryStickerPutResultModel : Codable {
     var isSuccess : Bool
     var code : Int
