+++ Todoary-iOS/Todoary/Network/Service/Auth/AccountService.swift
@@ -14,6 +14,18 @@ class AccountService: BaseService{
 
 extension AccountService {
     
+    func login(request: LoginInput, completion: @escaping (NetworkResult<Any>) -> Void){
+        requestObjectWithEmptyResponse(AccountRouter.login(requeset: request), completion: completion)
+    }
+    
+    func autoLogin(request: AutoLoginInput, completion: @escaping (NetworkResult<Any>) -> Void){
+        requestObjectWithEmptyResponse(AccountRouter.autoLogin(request: request), completion: completion)
+    }
+    
+    func logout(request: SignoutInput, completion: @escaping (NetworkResult<Any>) -> Void){
+        requestObjectWithEmptyResponse(AccountRouter.logout, completion: completion)
+    }
+    
     func generateAccount(request: SignUpRequestModel, completion: @escaping (NetworkResult<Any>) -> Void){
         requestObjectWithEmptyResponse(AccountRouter.signUp(request: request), completion: completion)
     }
@@ -30,7 +42,11 @@ extension AccountService {
         requestObjectWithEmptyResponse(AccountRouter.deleteAppleAccount(request: request), completion: completion)
     }
     
-    func getIsUserEmailDuplicate(email: String, completion: @escaping (NetworkResult<Any>) -> Void){
+    func checkUserEmailDuplicate(email: String, completion: @escaping (NetworkResult<Any>) -> Void){
         requestObjectWithEmptyResponse(AccountRouter.emailDuplicate(request: email), completion: completion)
+    }
+    
+    func modifyPassword(request: PwFindInput, completion: @escaping (NetworkResult<Any>) -> Void){
+        requestObjectWithEmptyResponse(AccountRouter.patchPassword(requeset: request), completion: completion)
     }
 }
