--- Todoary-iOS/Todoary/Network/Service/Auth/WithdrawalService.swift
@@ -1,8 +1,0 @@
-//
-//  WithdrawalService.swift
-//  Todoary
-//
-//  Created by 박지윤 on 2023/01/10.
-//
-
-import Foundation
