+++ Todoary-iOS/Todoary/Network/Service/Diary/DiaryRouter.swift
@@ -1,5 +1,5 @@
 //
-//  DiaryTextRouter.swift
+//  DiaryRouter.swift
 //  Todoary
 //
 //  Created by 박지윤 on 2023/01/10.
@@ -8,26 +8,24 @@
 import Foundation
 import Alamofire
 
-enum DiaryTextRouter{
+enum DiaryRouter{
     case postDiary(date: String, request: DiaryRequestModel)
     case deleteDiary(date: String)
     case getDiary(date: String)
+    case getDiarySticker
+    case putDiarySticker(id: Int, request: DiaryStickerInput)
 }
 
-/*
- path
- 
- POST       diary       "/diary"
- Delete     diary       "/diary"
- GET        diary       "/diary"
- */
-extension DiaryTextRouter: BaseRouter{
+extension DiaryRouter: BaseRouter{
     
     var path: String{
         switch self{
         case .postDiary(let date, _):           return HTTPMethodURL.POST.diary + "/\(date)"
         case .deleteDiary(let date):            return HTTPMethodURL.POST.diary + "/\(date)"
         case .getDiary:                         return HTTPMethodURL.GET.diary
+        case .getDiarySticker:                        return HTTPMethodURL.GET.sticker
+        case .putDiarySticker(let id, _):             return HTTPMethodURL.PUT.sticker + "/\(id)"+"/sticker"
+            //url 수정 필요
         }
     }
     
@@ -36,6 +34,8 @@ extension DiaryTextRouter: BaseRouter{
         case .postDiary:            return .post
         case .deleteDiary:          return .delete
         case .getDiary:             return .get
+        case .getDiarySticker:                        return .get
+        case .putDiarySticker:                         return .put
         }
     }
     
@@ -46,6 +46,8 @@ extension DiaryTextRouter: BaseRouter{
         case .getDiary(let date):
             let query: [String:Any] = ["createdDate" : date]
                                             return .query(query)
+        case .getDiarySticker:                        return .requestPlain
+        case .putDiarySticker(_, let request):        return .requestBody(request)
         }
     }
